// SGUEES-SPA/src/app/pages/Accounting/con-partida-modelo-deta/models/con-partida-modelo-deta.ts
export interface ConPartidaModeloDeta {
	CORR_EMPRESA: number;
	CORR_CLASE_PARTIDA: number;
	CORR_PARTIDA: number;
	CORR_PARTIDA_DETA: number;
	CUENTA_CONTABLE: string;
	NOMBRE_CUENTA: string;
	CORR_CENTRO_COSTO: number;
	NOMBRE_CENTRO: string;
	MONTO_CARGO: number;
	MONTO_ABONO: number;
	NOMBRE_TRAN: string;
}